requirements.txt
# core
discord.py>=2.0.0
aiohttp
aiosqlite

# utils
python-dateutil
psutil
beautifulsoup4
pytz

# media
Pillow
youtube_dl

# deprecate those
apiclient
wikipedia
wikidata
twitter
